backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic-settings",
    "sqlalchemy",
    "asyncpg",
    "greenlet",
    "pyjwt",
    "bcrypt",
    "python-multipart",
    "email-validator",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "aiosqlite",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
    "langgraph-cli[inmem]>=0.4.11",
    "langchain-groq>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "instagrapi",
    "requests",
]
